refactor: 바텀시트 파일 위치 변경 #37
